+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * ECS - Member consortial accordion is not displaying when a user has affiliations but does not have permission to view holdings. Fixes UIIN-2843.
 * Do not remove field from the form when its value is an empty string. Fixes UIIN-2787.
 * ECS: Shared instance cannot be edited from member tenant, even with permissions in both Central and member tenants. Fixes UIIN-2832.
+* Restricted status displays as Order Closed. Fixes UIIN-2821.
 
 ## [11.0.0](https://github.com/folio-org/ui-inventory/tree/v11.0.0) (2024-03-21)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v10.0.11...v11.0.0)